refactor(discovery): decompose the engine god-file by role (WS-D)
Split internal/discovery/engine.go (678 LOC) into role-based files in the
same package — pure move: engine types, config, constructors, collector
setters, and lifecycle stay in engine.go; the scan orchestration/phases
move out.

All files under the 600-line size-gate threshold. Declaration sets and
code-line multisets verified identical to the original.

Related to #187
